feat: Add GitHub links to portfolio projects
Introduces the `githubUrl` field to the project type definition. This allows for linking to the source code repositories of portfolio projects. The PortfolioModal component is updated to display a GitHub icon and link when a `githubUrl` is present. Additionally, the "Remittance Calculator BD" and a new "Alumni Association Website" project have their GitHub links added to the constants.